+++ octgnFX/Octgn/Windows/Options.xaml.cs
@@ -40,6 +40,7 @@
             CheckBoxUseInstantSearch.IsChecked = Prefs.InstantSearch;
             CheckBoxEnableGameSounds.IsChecked = Prefs.EnableGameSound;
             ComboBoxZoomOptions.SelectedIndex = (int)Prefs.ZoomOption;
+            ComboBoxJoinSound.SelectedIndex = (int)Prefs.SoundOption;
             CheckBoxEnableGameFonts.IsChecked = Prefs.UseGameFonts;
             CheckBoxEnableAdvancedOptions.IsChecked = Prefs.EnableAdvancedOptions;
             ComboBoxCardMoveNotification.SelectedIndex = (int)Prefs.CardMoveNotification;
@@ -130,6 +131,7 @@
             var handDensity = HandDensitySlider.Value;
             var useTestReleases = CheckBoxUseTestReleases.IsChecked ?? false;
             Prefs.ZoomType zoomOption = (Prefs.ZoomType)ComboBoxZoomOptions.SelectedIndex;
+            Prefs.SoundType soundOption = (Prefs.SoundType)ComboBoxJoinSound.SelectedIndex;
             Prefs.CardAnimType animOption = (Prefs.CardAnimType)ComboBoxCardMoveNotification.SelectedIndex;
             var task = new Task(
                 () => 
@@ -150,7 +152,10 @@
                     chatFontSize,
                     useInstantSearch,
                     enableGameSounds,
-                    zoomOption,animOption,enableAdvancedOptions,
+                    zoomOption,
+                    soundOption,
+                    animOption,
+                    enableAdvancedOptions,
                     useGameFonts,
                     handDensity,
                     useTestReleases)
@@ -182,6 +187,7 @@
             bool useInstantSearch,
             bool enableGameSounds,
             Prefs.ZoomType zoomOption,
+            Prefs.SoundType soundOption,
             Prefs.CardAnimType animOption,
             bool enableAdvancedOptions,
             bool useGameFonts,
@@ -226,6 +232,7 @@
             Prefs.ChatFontSize = chatFontSize;
             Prefs.InstantSearch = useInstantSearch;
             Prefs.EnableGameSound = enableGameSounds;
+            Prefs.SoundOption = soundOption;
             Prefs.ZoomOption = zoomOption;
             Prefs.CardMoveNotification = animOption;
             Prefs.EnableAdvancedOptions = enableAdvancedOptions;